move LoopbackCallbackResult to Euterpe.Contracts.Account
