+++ src/Euterpe.Contracts/Account/LoopbackCallbackResult.cs
@@ -1,3 +1,4 @@
-namespace Euterpe.Abstractions;
+namespace Euterpe.Contracts.Account;
 
+[PublicAPI]
 public sealed record LoopbackCallbackResult(string? Code, string? State, string? Error);
